chore: move changes into stream-ui for clearer diff
