--- examples/next-openai/app/stream-ui-nested/layout.tsx
@@ -1,5 +1,0 @@
-import { AI } from './actions';
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <AI>{children}</AI>;
-}
