+++ examples/next-openai/app/stream-ui/page.tsx
@@ -13,16 +13,16 @@ export default function Home() {
   const [messages, setMessages] = useUIState<typeof AI>();
   const { submitUserMessage } = useActions<typeof AI>();
 
-  const handleSubmission = async () => {
+  const handleSubmission = async (message: string) => {
     setMessages(currentMessages => [
       ...currentMessages,
       {
         id: generateId(),
-        display: <Message role="user">{input}</Message>,
+        display: <Message role="user">{message}</Message>,
       },
     ]);
 
-    const response = await submitUserMessage(input);
+    const response = await submitUserMessage(message);
     setMessages(currentMessages => [...currentMessages, response]);
     setInput('');
   };
@@ -37,19 +37,33 @@ export default function Home() {
           placeholder="Ask a question"
           onKeyDown={event => {
             if (event.key === 'Enter') {
-              handleSubmission();
+              handleSubmission(input);
             }
           }}
         />
         <button
           className="p-2 bg-zinc-900 text-zinc-100 rounded-md"
-          onClick={handleSubmission}
+          onClick={() => handleSubmission(input)}
         >
           Send
         </button>
       </div>
 
       <div className="flex flex-col h-[calc(100dvh-56px)] overflow-y-scroll">
+        {messages.length === 0 && (
+          <div className="flex items-center justify-center size-full">
+            <button
+              className="px-4 py-2 bg-zinc-900 text-zinc-100 rounded-md w-fit"
+              onClick={() => {
+                handleSubmission(
+                  'Show me the weather for San Francisco, New York, and Chicago',
+                );
+              }}
+            >
+              Test With Example Prompt
+            </button>
+          </div>
+        )}
         <div>
           {messages.map(message => (
             <Fragment key={message.id}>{message.display}</Fragment>
